% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answer_as_binary.R
\name{answer_as_binary}
\alias{answer_as_binary}
\title{Yes/No answer to binary}
\usage{
answer_as_binary(df, vars)
}
\arguments{
\item{df}{The dataframe containing the information.}

\item{vars}{a variable (or vector of variables) quoted that you want to change to binary.}
}
\value{
The binary answers, changing "no" or "none" to 0 and "yes" to 1.
}
\description{
This function allows you to convert multiple yes/no answers in binary format
}
\examples{
answer_as_binary(df, "tenencia_bici")
}
\keyword{answer}
\keyword{binary}
\keyword{survey}
