% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colores_EMU.R
\name{colores_EMU}
\alias{colores_EMU}
\title{A function that retreives specific colors for EMU}
\usage{
colores_EMU()
}
\arguments{
\item{NULL}{It doesn't require parameters}
}
\value{
The four color vectors
}
\description{
This function creates 4 vectors with categorical colors for the EMU
}
\examples{
colores_EMU()
}
\keyword{EMU}
\keyword{colors}
\keyword{palette}
