% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emu_tabyl.R
\name{emu_tabyl}
\alias{emu_tabyl}
\title{A function to create weighted crosstables}
\usage{
emu_tabyl(df, var1, var2, var3, expansion = f_exp, show_na = T, ...)
}
\arguments{
\item{df}{The dataframe containing the EMU information.}

\item{var1}{the column name of the first variable.}

\item{var2}{(optional) the column name of the second variable (the rows in a 2-way tabulation).}

\item{var3}{(optional) the column name of the third variable (the list in a 3-way tabulation).}

\item{expansion}{the column name with the expansion factor to be used in the summarise}

\item{show_na}{should counts of NA values be displayed? In a one-way tabyl, the presence of NA values triggers an additional column showing valid percentages(calculated excluding NA values).}
}
\value{
The pivot table with the expansion variable.
}
\description{
This function allows you to create pivot tables for EMU, expanded and non expanded.
}
\examples{
emu_tabyl(bd, Antes_viajes_2_6, show_na = F)
}
\keyword{expansion}
\keyword{factor}
\keyword{pivot}
\keyword{table}
