% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factores_EMU.R
\name{factores_EMU}
\alias{factores_EMU}
\title{A function to calculate expansions factors in EMU}
\usage{
factores_EMU(df, as.df)
}
\arguments{
\item{df}{The dataframe containing the depurated EMU information}

\item{as.df}{If true, returns a dataframe with the information for the expansion factors}

\item{as.data}{If true, returns only the values of the expansion factors}
}
\value{
The expansion factors
}
\description{
This function allows you calculate expansion factors per poblation in EMU, with a dataframe containing population
}
\examples{
factores_EMU(df, as.df, as.data)
}
\keyword{EMU}
\keyword{expansion}
\keyword{factors}
\keyword{survey}
