% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_Colombia.R
\name{geocode_Colombia}
\alias{geocode_Colombia}
\title{geocode in Colombia}
\usage{
geocode_Colombia(df, CRS, id, address, city, names_crs, names_sep)
}
\arguments{
\item{df}{The dataframe containing the information.}

\item{CRS}{The CRS for geocoding the addresses.}

\item{id}{The name of the column with the id for keeping track of your data.}

\item{address}{The name of the column with the addresses for geocoding.}

\item{city}{The name of the column with the city of the address.}
}
\value{
The coordinates of the address given, and the match score.
}
\description{
This function allows you to geocode addresses in Colombia, specifying address, city and a CRS.
}
\examples{
geocode_Colombia(df, 3116, "id","street", "city")
}
\keyword{Colombia}
\keyword{address.}
\keyword{geocode}
