% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocodeAddresses.R
\name{geocodeAddresses}
\alias{geocodeAddresses}
\title{Geocode addresses in Colombia}
\usage{
geocodeAddresses(
  df,
  address,
  city,
  countryCode = "CO",
  crs = 3116,
  names_crs,
  names_sep
)
}
\arguments{
\item{df}{The dataframe containing the information.}

\item{address}{The name of the column with the addresses for geocoding.}

\item{city}{The name of the column with the city of the address.}

\item{countryCode}{Limits the candidates returned to the specified country or countries. Acceptable values include the 3-character country code. You can specify multiple country codes to limit results to more than one country.}

\item{crs}{The spatial reference of the x/y coordinates returned by a geocode request. The spatial reference can be specified as either a well-known ID (WKID) or as a JSON spatial reference object. If crs is not specified, the spatial reference of the output locations is 3116}

\item{names_crs}{Logical. If the crs must be added at the end of the results columns "lon" and "lat"}

\item{names_sep}{A string to separate the name of the columns "lon" and "lat" and the crs (WKID)}
}
\value{
The coordinates (lon, lat) of the address given and the match score.
}
\description{
This function allows you to geocode addresses in Colombia, specifying address, city and a CRS.
}
\examples{
df \%>\% geocodeAddresses(address = direccion_hogar, city = ciudad)
}
\keyword{Colombia}
\keyword{address.}
\keyword{geocode}
