% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsave2.R
\name{ggsave2}
\alias{ggsave2}
\title{A function that allows piping ggsave}
\usage{
ggsave2(...)
}
\arguments{
\item{NULL}{It doesn't require parameters}
}
\value{
A saved ggplot object
}
\description{
A function that allows piping ggsave
}
\examples{
ggsave2()
}
\keyword{ggsave}
