% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_addresses.R
\name{paste_addresses}
\alias{paste_addresses}
\title{A function to paste the input for street in geocode_Colombia}
\usage{
paste_addresses(BD)
}
\arguments{
\item{df}{The dataframe containing the informaiton.}
}
\value{
The address.
}
\description{
This function allows you to paste the information from "Calle" and "Carrera" as an input for geocode_Colombia
}
\examples{
paste_addresses(df)
}
\keyword{Colombia}
\keyword{geocode}
\keyword{geocode_Colombia}
