% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_abs.R
\name{percent_abs}
\alias{percent_abs}
\title{Labels in format percent and absolute for ggplot graphs}
\usage{
percent_abs(x, ...)
}
\arguments{
\item{x}{The labels.}

\item{...}{Another aesthetical value.}
}
\value{
Labs in percent and absolute format for ggplot graphics.
}
\description{
This function gives a labels format based on percentage and absolute values.
}
\examples{
labels = percent_abs()
}
\keyword{absolute}
\keyword{ggplot}
\keyword{labs}
\keyword{percent}
