% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_table_EMU.R
\name{pivot_table_EMU}
\alias{pivot_table_EMU}
\title{A function to create pivot tables}
\usage{
pivot_table_EMU(df, vars, expanded, digits.expanded, digits.percentage)
}
\arguments{
\item{df}{The dataframe containing the EMU information.}

\item{vars}{a variable (or vector of variables) quoted that you want use as input for pivot table.}

\item{expanded}{T if expanded, F if unexpanded. If missins, asummes TRUE.}
}
\value{
The pivot table.
}
\description{
This function allows you to create pivot tables for EMU, expanded and non expanded.
}
\examples{
pivot_table_EMU(df, "hacia_modo_ppal", TRUE)
}
\keyword{EMU}
\keyword{pivot}
\keyword{table}
