% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_emu.R
\name{quantile_emu}
\alias{quantile_emu}
\title{A function to calculate the quantile from a numeric variable}
\usage{
quantile_emu(
  df,
  var,
  min.seq,
  max.seq,
  by,
  labels,
  labels.mark,
  include.lowest
)
}
\arguments{
\item{df}{The dataframe containing the EMU information.}

\item{var}{The column name of the variable to calculate the quantile position.}

\item{min.seq}{The minimum value to compute the probablities of the quantile (default to 0)}

\item{max.seq}{The maximum value to compute the probablities of the quantile (default to 1)}

\item{by}{The interval to compute the probabilities calculation (default to 0.25 for 4 quartiles)}

\item{labels}{Logical. If labels should be calculated, instead of the brackets position}

\item{labels.mark}{Interval. Labels for the quartiles calculated depending on the intervals used as inputs (default 1:4)}

\item{include.lowest}{Logical. If lowest value from the interval should be included}
}
\value{
The column with the quartile computed
}
\description{
This function allows you to calculate the quantile position from a variable in a dataframe
}
\examples{
quantile_emu(bd, var = value)
}
\keyword{quartile}
