% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_as_mean.R
\name{range_as_mean}
\alias{range_as_mean}
\title{A function to convert a range of values into their mean}
\usage{
range_as_mean(df, vars)
}
\arguments{
\item{df}{The dataframe containing the variables you want to changes.}

\item{vars}{a variable (or vector of variables) quoted that you want to change from a range to their mean.}
}
\value{
The mean of \code{vars}.
}
\description{
This function allows you to convert each range into their mean, even if it has string elements, make sure your
delete any don't wanted numbers in the range.
}
\examples{
range_as_mean(df, "salary")
}
\keyword{mean}
\keyword{range}
