% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_code.R
\name{save_code}
\alias{save_code}
\title{Save Apollo code and rename older ones}
\usage{
save_code(apollo_control)
}
\arguments{
\item{apollo_control}{The objetct with the apollo_control information}
}
\value{
The code in a R file
}
\description{
This function allows you to save Apollo code and rename older ones
}
\examples{
save_code(apollo_control)
}
\keyword{Apollo,}
\keyword{code}
\keyword{save}
