% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_image.R
\name{save_image}
\alias{save_image}
\title{Save Apollo image space and rename older ones}
\usage{
save_image(apollo_control)
}
\arguments{
\item{apollo_control}{The objetct with the apollo_control information}
}
\value{
The workspace in a RData file
}
\description{
This function allows you to save Apollo image spaces and rename older ones
}
\examples{
save_image(apollo_control)
}
\keyword{gApollo,}
\keyword{image}
\keyword{save}
