% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_to_dodger.R
\name{sf_to_dodgr}
\alias{sf_to_dodgr}
\title{A function that transform a sf object into a dodgr object}
\usage{
sf_to_dodgr(net, id_col, wt_profile)
}
\arguments{
\item{net}{The sf object}

\item{id_col}{String character with the name of the column with the id of the network}

\item{wt_profile}{a json object or custom weight profile (see dodgr package), default is 1}
}
\value{
A dodgr object
}
\description{
This function return a dodger object (as a dataframe) from a sf object. Is usefull for net_dist and net_path
}
\examples{

sf_to_dodger(tm)
}
\keyword{dodgr}
\keyword{sf}
