% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_round_edges.R
\name{sm_round_edges}
\alias{sm_round_edges}
\title{Round edges for ggplot graphs}
\usage{
sm_round_edges(plot = NULL, radio = NULL)
}
\arguments{
\item{plot}{The resulting ggplot graph.}

\item{radio}{The radio in percentage for rounding the edges.}
}
\value{
The plot with rounded edges.
}
\description{
This function allows you to round the edges of a ggplot graph.
}
\examples{
geocode_Colombia(plot, r = 0.05)
}
\keyword{edges}
\keyword{ggplot}
\keyword{round}
