% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_steer.R
\docType{data}
\name{steer_colours}
\alias{steer_colours}
\title{A function to use the Steer colors in ggplot graphs}
\format{
An object of class \code{list} of length 1.
}
\usage{
steer_colours
}
\arguments{
\item{name}{The name of the palette}

\item{direction}{The direction for using the colors}
}
\value{
The Steer palette for ggplot graphs
}
\description{
This function allows you to geocode addresses in Colombia and get the results as a readable dataframe.
}
\examples{
scale_fill_steer_d(name = "steer", direction = -1,)
}
\keyword{Steer}
\keyword{colours}
\keyword{palette}
