% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_emu.R
\name{theme_emu}
\alias{theme_emu}
\title{Theme for EMU graphs}
\usage{
theme_emu(
  base_size = 11,
  base_family = NULL,
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{Size for letters.}

\item{base_family}{Font used for the elements on the plot, different from those inserted with geom_text.}

\item{base_line_size}{Line size}

\item{base_rect_size}{Base rect size}
}
\value{
A theme for ggplot graphics.
}
\description{
This function gives a nice looking theme for ggplot graphs.
}
\examples{
theme(base_size = 30, base_family = "Gotham Rounded Book")
}
\keyword{emu}
\keyword{ggplot}
\keyword{theme}
