% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_google_Colombia.R
\name{geocode_google_Colombia}
\alias{geocode_google_Colombia}
\title{A function to geocode colombian addresses using google geocoding API and googleway package}
\usage{
geocode_google_Colombia(
  df,
  direccion,
  key,
  suffix,
  barrio,
  codigo_postal,
  join,
  join_id,
  ...
)
}
\arguments{
\item{df}{The df containing the addresses}

\item{direccion}{The name of the column (quoted) where the address is located (it must be formed as "address", "city", "country")}

\item{key}{The google API key used for geocoding}

\item{suffix}{The suffix for the resultant columns generated by the geocoding}

\item{barrio}{Logical. If true, returns the information of the neighborhood}

\item{codigo_postal}{Logical. If true, returns the information of the postal code}

\item{join}{Logical. If true, the resultant dataframe is united with the original dataframe}

\item{join_id}{A string indicating the name of the column to join the results into the new dataframe}
}
\value{
A dataframe containing the coordinates in crs 4326 with the addresses geocoded
}
\description{
This function allows you to geocode addresses in Colombia and get the results as a readable dataframe.
}
\examples{
direcciones <- geocode_google_Colombia(df, "direccion_google", key = "ABC", suffix = "_com", barrio = T, codigo_postal = T)
}
\keyword{Bogota}
\keyword{Colombia}
\keyword{geocode}
\keyword{google}
