% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_multiple_Colombia.R
\name{geocode_multiple_Colombia}
\alias{geocode_multiple_Colombia}
\title{geocode in Colombia using the ESRI API geocoding server first}
\usage{
geocode_multiple_Colombia(df, key, id_col = NULL, join_id = NULL, ...)
}
\arguments{
\item{df}{The dataframe containing the information.}

\item{key}{The google API key used for geocoding with the Google Maps API.}

\item{id_col}{The name of the column with the id for keeping track of your data (for each address).}

\item{join_id}{A string indicating the name of the column to join the results into the new dataframe.}
}
\value{
The coordinates of the address given.
}
\description{
This function allows you to geocode addresses in Colombia. The first attempt to geocode addresses is by using the ESRI
API geocoding server. If any address is not geocoded, then this directions are geocoded using the Google Maps API
(requires key). The result is a datframe with all the information recolected, specifying in the column geocoded with which
server the address was geocoded (1: ESRI, 2: Google Maps). A column for localidad is specified when the address is geocoded
with the Google Maps API, since the score of this geocoding process is not computated, so this may be the fastes way to
verify the results obtained.
}
\examples{
ver <- geocode_multiple(bd, key)
}
\keyword{Colombia,}
\keyword{Google}
\keyword{Maps.}
\keyword{address,}
\keyword{geocode,}
\keyword{geocoding,}
